requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
greenlet==3.2.4
idna==3.11
kiwisolver==1.4.9
llvmlite==0.45.1
matplotlib==3.10.7
numba==0.62.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
playwright==1.55.0
pyee==13.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
typing-extensions==4.15.0
urllib3==2.5.0
